Extract logic for retrieving a single recipe based on recipeId to separate sql-function
